+++ src/components/subscription/index.ts
@@ -2,3 +2,4 @@
 
 export { FeatureGate, FeatureButtonGate, withFeatureGate } from "./FeatureGate";
 export { UpgradePrompt, UpgradePromptCompact } from "./UpgradePrompt";
+export { AnalyticsSectionGate } from "./AnalyticsSectionGate";